fix: cast numeric params in payment-status UPDATE (500 on every payment)
The overpayment-guard commit (0505dc2) left amount_paid params untyped in
the UPDATE, so Postgres threw 'operator is not unique: unknown + unknown'
on every payment. Cast \/\ to numeric. Verified end-to-end against the
real route code: partial pay 200, overpay 409, exact remainder 200.

+++ backend/src/routes/billingRoutes.js
@@ -1248,14 +1248,14 @@ router.post('/invoice-payments', auth, async (req, res) => {
 
     await client.query(`
       UPDATE invoices
-      SET amount_paid = $1,
+      SET amount_paid = $1::numeric,
           payment_status = CASE
-            WHEN $1 + $2 >= total THEN 'paid'
-            WHEN $1 > 0 THEN 'partial'
+            WHEN $1::numeric + $2::numeric >= total THEN 'paid'
+            WHEN $1::numeric > 0 THEN 'partial'
             ELSE 'pending'
           END,
-          payment_date = CASE WHEN $1 + $2 >= total THEN $3 ELSE payment_date END,
-          paid_at = CASE WHEN $1 + $2 >= total THEN NOW() ELSE paid_at END,
+          payment_date = CASE WHEN $1::numeric + $2::numeric >= total THEN $3 ELSE payment_date END,
+          paid_at = CASE WHEN $1::numeric + $2::numeric >= total THEN NOW() ELSE paid_at END,
           updated_at = NOW()
       WHERE id = $4
     `, [newPaid, adjusted, paymentDate, invoiceId]);
